Updated AWESim_SOSS notebook, renamed synSOSS to sim2D to stick with naming convention, and added extract1D notebook.
